+++ packages/maestro/lib/maestro.dart
@@ -6,4 +6,5 @@ export 'package:maestro/src/inspector.dart';
 export 'package:maestro/src/listener.dart';
 export 'package:maestro/src/maestro.dart';
 export 'package:maestro/src/maestros.dart';
+export 'package:maestro/src/memento.dart';
 export 'package:maestro/src/selector.dart';